Reorganize config (#157)
* Reorganize config

* allow missing docs so tests run

* lint abatement

* some renames
