+++ src/append/rolling_file/policy/compound/roll/delete.rs
@@ -8,7 +8,7 @@ use failure::Error;
 
 use crate::append::rolling_file::policy::compound::roll::Roll;
 #[cfg(feature = "config_parsing")]
-use crate::config_parsing::{Deserialize, Deserializers};
+use crate::config::{Deserialize, Deserializers};
 
 /// Configuration for the delete roller.
 #[cfg(feature = "config_parsing")]
