+++ src/append/rolling_file/policy/compound/trigger/mod.rs
@@ -6,7 +6,7 @@ use failure::Error;
 
 use crate::append::rolling_file::LogFile;
 #[cfg(feature = "config_parsing")]
-use crate::config_parsing::Deserializable;
+use crate::config::Deserializable;
 
 #[cfg(feature = "size_trigger")]
 pub mod size;
