+++ src/lib.rs
@@ -185,39 +185,27 @@
 
 #![allow(where_clauses_object_safety, clippy::manual_non_exhaustive)]
 #![warn(missing_docs)]
+// TODO: need to remove before merge
+#![allow(missing_docs, clippy::module_inception)]
 
 use arc_swap::ArcSwap;
 use fnv::FnvHasher;
-use log::{Level, LevelFilter, Metadata, Record, SetLoggerError};
+use log::{Level, LevelFilter, Metadata, Record};
 use std::{cmp, collections::HashMap, hash::BuildHasherDefault, io, io::prelude::*, sync::Arc};
-
-use failure::{Error, Fail};
-
-#[cfg(feature = "config_parsing")]
-pub use crate::priv_file::{init_file, load_config_file, FormatError};
-
-#[cfg(feature = "config_parsing")]
-pub use crate::config_parsing::{Deserializers, RawConfig};
-
-use crate::{append::Append, config::Config, filter::Filter};
 
 pub mod append;
 pub mod config;
 pub mod encode;
 pub mod filter;
-
-#[cfg(feature = "config_parsing")]
-pub mod config_parsing;
-#[cfg(feature = "config_parsing")]
-mod priv_file;
-
 #[cfg(feature = "console_writer")]
 mod priv_io;
 
-/// Collects the set of errors that occur when deserializing the appenders.
-#[derive(Debug, Fail)]
-#[fail(display = "Errors on initialization: {:#?}", _0)]
-pub struct InitErrors(Vec<Error>);
+pub use config::{init_config, Config};
+
+#[cfg(feature = "config_parsing")]
+pub use config::{init_file, init_raw_config};
+
+use self::{append::Append, filter::Filter};
 
 type FnvHashMap<K, V> = HashMap<K, V, BuildHasherDefault<FnvHasher>>;
 
@@ -325,7 +313,6 @@ struct SharedLogger {
     root: ConfiguredLogger,
     appenders: Vec<Appender>,
 }
-
 impl SharedLogger {
     fn new(config: config::Config) -> SharedLogger {
         let (appenders, root, mut loggers) = config.unpack();
@@ -413,43 +400,9 @@ impl log::Log for Logger {
     }
 }
 
-fn handle_error(e: &failure::Error) {
+pub(crate) fn handle_error(e: &failure::Error) {
     let _ = writeln!(io::stderr(), "log4rs: {}", e);
 }
-
-/// Initializes the global logger as a log4rs logger with the provided config.
-///
-/// A `Handle` object is returned which can be used to adjust the logging
-/// configuration.
-pub fn init_config(config: config::Config) -> Result<Handle, SetLoggerError> {
-    let logger = Logger::new(config);
-    log::set_max_level(logger.max_log_level());
-    let handle = Handle {
-        shared: logger.0.clone(),
-    };
-    log::set_boxed_logger(Box::new(logger)).map(|()| handle)
-}
-
-/// Initializes the global logger as a log4rs logger using the provided raw config.
-///
-/// This will return errors if the appenders configuration is malformed or if we fail to set the global logger.
-#[cfg(feature = "config_parsing")]
-pub fn init_raw_config(config: RawConfig) -> Result<(), Error> {
-    let (appenders, errors) = config.appenders_lossy(&Deserializers::default());
-    if !errors.is_empty() {
-        return Err(InitErrors(errors).into());
-    }
-    let config = Config::builder()
-        .appenders(appenders)
-        .loggers(config.loggers())
-        .build(config.root())?;
-
-    let logger = Logger::new(config);
-    log::set_max_level(log::LevelFilter::Info);
-    log::set_boxed_logger(Box::new(logger))?;
-    Ok(())
-}
-
 /// A handle to the active logger.
 #[derive(Clone)]
 pub struct Handle {
@@ -467,14 +420,6 @@ impl Handle {
 
 trait ErrorInternals {
     fn new(message: String) -> Self;
-}
-
-trait ConfigPrivateExt {
-    fn unpack(self) -> (Vec<config::Appender>, config::Root, Vec<config::Logger>);
-}
-
-trait PrivateConfigAppenderExt {
-    fn unpack(self) -> (String, Box<dyn Append>, Vec<Box<dyn Filter>>);
 }
 
 #[cfg(test)]
@@ -505,7 +450,7 @@ mod test {
                 }
             },
         });
-        let config = serde_json::from_str::<config_parsing::RawConfig>(&cfg.to_string()).unwrap();
+        let config = serde_json::from_str::<config::RawConfig>(&cfg.to_string()).unwrap();
         if let Err(e) = init_raw_config(config) {
             panic!(e);
         }
